consecutive_rise_finder/requirements.txt
# LS증권 연속 상승 탐색기 의존성
requests>=2.28.0
python-dotenv>=1.0.0
